+++ constants.go
@@ -1,0 +1,9 @@
+package main
+
+// service related information to deliver more context to the logs
+const (
+	serviceGroup   = "jsa"
+	serviceName    = "grpc-file-streamer"
+	serviceType    = "service"
+	serviceVersion = "0.1.0"
+)