+++ logging/messages.go
@@ -1,0 +1,4 @@
+package logging
+
+// Common Info Messages
+const ()